further improvements

+++ resources/assets/js/app/messageHandling.js
@@ -7,21 +7,22 @@ export default {
 
   getAllMessages () {
     // get all existing messages
-    axios.get('/messages').then((response) => {
-      this.messages = response.data;
-    });
+    axios.get('/messages')
+      .then((response) => {
+        this.messages = response.data;
+      });
   },
 
   addMessage (payload) {
+    // send the new message to the backend
     axios.post('/messages', { message: payload})
       .then((response) => {
+        // once the backend processed the new message, it will be broadcasted
         // console.log(response);
-        // this.getAllMessages();
       })
   },
 
   editMessage (payload) {
-    console.log('editMessage method: '+payload.message);    
     axios.patch('/messages/'+payload.id, payload)
       .then((response) => {
         if (response.data == 'updated')
@@ -32,11 +33,15 @@ export default {
   },
 
   deleteMessage (payload) {
+    if (parseInt(payload) == isNaN)
+      return;
     axios.delete('/messages/' + payload)
       .then((response) => {
         if (response.data == 'destroyed') {
           this.messages = this.messages.filter( (elem) => { return elem.id != payload } );
         }
+        else 
+          console.log(response);
       })
   },
